# go/docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Sports analytics API serving player/team profiles, stats, percentiles, news, and journalist tweets. All data-heavy responses are JSON-passthrough from Postgres functions.",
        "title": "Scoracle Data API",
        "contact": {
            "name": "Scoracle"
        },
        "license": {
            "name": "MIT"
        },
        "version": "2.0.0"
    },
    "host": "localhost:8000",
    "basePath": "/api/v1",
    "paths": {
...
        "/stats/definitions": {
            "get": {
                "description": "Returns canonical stat definitions (names, display labels, categories) for a sport.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get stat definitions",
                "parameters": [
                    {
                        "enum": [
                            "NBA",
                            "NFL",
                            "FOOTBALL"
                        ],
                        "type": "string",
                        "description": "Sport identifier",
                        "name": "sport",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_albapepper_scoracle-data_go_internal_api_respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_albapepper_scoracle-data_go_internal_api_respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stats/{entityType}/{entityID}": {
            "get": {
                "description": "Returns stats and percentiles for a player or team for a given season. Response is raw JSON from Postgres api_entity_stats function.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get entity stats",
                "parameters": [
                    {
                        "enum": [
                            "player",
                            "team"
                        ],
                        "type": "string",
                        "description": "Entity type",
                        "name": "entityType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Entity ID",
                        "name": "entityID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "NBA",
                            "NFL",
                            "FOOTBALL"
                        ],
                        "type": "string",
                        "description": "Sport identifier",
                        "name": "sport",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Season year (defaults to current)",
                        "name": "season",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "League ID (for FOOTBALL)",
                        "name": "league_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_albapepper_scoracle-data_go_internal_api_respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_albapepper_scoracle-data_go_internal_api_respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/stats/{entityType}/{entityID}/seasons": {
            "get": {
                "description": "Returns a list of seasons that have stats for an entity.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "stats"
                ],
                "summary": "Get available seasons",
                "parameters": [
                    {
                        "enum": [
                            "player",
                            "team"
                        ],
                        "type": "string",
                        "description": "Entity type",
                        "name": "entityType",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "Entity ID",
                        "name": "entityID",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "NBA",
                            "NFL",
                            "FOOTBALL"
                        ],
                        "type": "string",
                        "description": "Sport identifier",
                        "name": "sport",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_albapepper_scoracle-data_go_internal_api_respond.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_albapepper_scoracle-data_go_internal_api_respond.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/twitter/journalist-feed": {
            "get": {
                "description": "Searches the cached journalist X List feed for entity mentions. Full feed is fetched once and cached for 1 hour; searches filter the cache client-side.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "twitter"
                ],
                "summary": "Search journalist feed",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query (1-200 chars)",
                        "name": "q",
                        "in": "query",
                        "required": true
                    },
                    {
                        "enum": [
                            "NBA",
                            "NFL",
                            "FOOTBALL"
                        ],
                        "type": "string",
                        "description": "Sport context (metadata only)",
                        "name": "sport",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Max tweets (1-50, default 10)",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_albapepper_scoracle-data_go_internal_api_respond.ErrorResponse"
                        }
                    },
                    "502": {
                        "description": "Bad Gateway",
                        "schema": {
                            "$ref": "#/definitions/github_com_albapepper_scoracle-data_go_internal_api_respond.ErrorResponse"
                        }
                    },
                    "503": {
                        "description": "Service Unavailable",
                        "schema": {
                            "$ref": "#/definitions/github_com_albapepper_scoracle-data_go_internal_api_respond.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_albapepper_scoracle-data_go_internal_api_respond.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "object",
                    "properties": {
                        "code": {
                            "type": "string"
                        },
                        "detail": {
                            "type": "string"
                        },
                        "message": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}